refactor: extract common code on redis unit test
