+++ support/cas-server-support-redis-service-registry/src/test/java/org/apereo/cas/adaptors/redis/services/RedisEmbeddedServiceRegistryTests.java
@@ -1,21 +1,10 @@
 package org.apereo.cas.adaptors.redis.services;
 
-import org.apereo.cas.config.RedisServiceRegistryConfiguration;
-import org.apereo.cas.configuration.CasConfigurationProperties;
-import org.apereo.cas.services.AbstractServiceRegistryTests;
-import org.apereo.cas.services.ServiceRegistry;
 import org.apereo.cas.util.junit.DisabledIfContinuousIntegration;
 
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
-import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.test.context.TestPropertySource;
 import redis.embedded.RedisServer;
 
@@ -25,18 +14,13 @@ import redis.embedded.RedisServer;
  * @author Misagh Moayyed
  * @since 4.0.0
  */
-@SpringBootTest(classes = {RedisServiceRegistryConfiguration.class, RefreshAutoConfiguration.class})
-@EnableScheduling
-@TestPropertySource(properties = {"cas.serviceRegistry.redis.host=localhost", "cas.serviceRegistry.redis.port=6380"})
-@EnableConfigurationProperties(CasConfigurationProperties.class)
-@Tag("Redis")
+@TestPropertySource(properties = {
+    "cas.serviceRegistry.redis.host=localhost",
+    "cas.serviceRegistry.redis.port=6380"
+})
 @DisabledIfContinuousIntegration
-public class RedisEmbeddedServiceRegistryTests extends AbstractServiceRegistryTests {
+public class RedisEmbeddedServiceRegistryTests extends BaseRedisSentinelServiceRegistryTests {
     private static RedisServer REDIS_SERVER;
-
-    @Autowired
-    @Qualifier("redisServiceRegistry")
-    private ServiceRegistry dao;
 
     @BeforeAll
     @SneakyThrows
@@ -50,8 +34,4 @@ public class RedisEmbeddedServiceRegistryTests extends AbstractServiceRegistryTe
         REDIS_SERVER.stop();
     }
 
-    @Override
-    public ServiceRegistry getNewServiceRegistry() {
-        return this.dao;
-    }
 }
