+++ support/cas-server-support-redis-ticket-registry/src/test/java/org/apereo/cas/ticket/registry/RedisEmbeddedTicketRegistryTests.java
@@ -1,23 +1,10 @@
 package org.apereo.cas.ticket.registry;
 
-import org.apereo.cas.config.CasCoreTicketCatalogConfiguration;
-import org.apereo.cas.config.CasCoreTicketsConfiguration;
-import org.apereo.cas.config.CasCoreWebConfiguration;
-import org.apereo.cas.config.RedisTicketRegistryConfiguration;
-import org.apereo.cas.config.support.CasWebApplicationServiceFactoryConfiguration;
 import org.apereo.cas.util.junit.DisabledIfContinuousIntegration;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.aop.AopAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.test.context.TestPropertySource;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
 import redis.embedded.RedisServer;
 
 /**
@@ -26,30 +13,15 @@ import redis.embedded.RedisServer;
  * @author Misagh Moayyed
  * @since 5.0.0
  */
-@Tag("Redis")
-@SpringBootTest(classes = {
-    RedisTicketRegistryConfiguration.class,
-    RefreshAutoConfiguration.class,
-    CasCoreWebConfiguration.class,
-    AopAutoConfiguration.class,
-    CasCoreTicketsConfiguration.class,
-    CasCoreTicketCatalogConfiguration.class,
-    CasWebApplicationServiceFactoryConfiguration.class})
 @TestPropertySource(properties = {
     "cas.ticket.registry.redis.host=localhost",
     "cas.ticket.registry.redis.port=6378",
     "cas.ticket.registry.redis.pool.max-active=20"
 })
-@EnableTransactionManagement(proxyTargetClass = true)
-@EnableAspectJAutoProxy(proxyTargetClass = true)
 @DisabledIfContinuousIntegration
-public class RedisEmbeddedTicketRegistryTests extends BaseTicketRegistryTests {
+public class RedisEmbeddedTicketRegistryTests extends BaseRedisSentinelTicketRegistryTests {
 
     private static RedisServer REDIS_SERVER;
-
-    @Autowired
-    @Qualifier("ticketRegistry")
-    private TicketRegistry ticketRegistry;
 
     @BeforeAll
     public static void startRedis() throws Exception {
@@ -62,8 +34,4 @@ public class RedisEmbeddedTicketRegistryTests extends BaseTicketRegistryTests {
         REDIS_SERVER.stop();
     }
 
-    @Override
-    public TicketRegistry getNewTicketRegistry() {
-        return this.ticketRegistry;
-    }
 }
