+++ support/cas-server-support-redis-ticket-registry/src/test/java/org/apereo/cas/ticket/registry/SentinelEmbeddedTicketRegistryTests.java
@@ -1,22 +1,10 @@
 package org.apereo.cas.ticket.registry;
 
-import org.apereo.cas.config.CasCoreTicketCatalogConfiguration;
-import org.apereo.cas.config.CasCoreTicketsConfiguration;
-import org.apereo.cas.config.CasCoreWebConfiguration;
-import org.apereo.cas.config.RedisTicketRegistryConfiguration;
-import org.apereo.cas.config.support.CasWebApplicationServiceFactoryConfiguration;
 import org.apereo.cas.util.junit.DisabledIfContinuousIntegration;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.aop.AopAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.test.context.TestPropertySource;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
 import redis.embedded.RedisSentinel;
 import redis.embedded.RedisServer;
 
@@ -26,15 +14,6 @@ import redis.embedded.RedisServer;
  * @author Julien Gribonvald
  * @since 6.1.0
  */
-@Tag("Redis")
-@SpringBootTest(classes = {
-    RedisTicketRegistryConfiguration.class,
-    RefreshAutoConfiguration.class,
-    CasCoreWebConfiguration.class,
-    AopAutoConfiguration.class,
-    CasCoreTicketsConfiguration.class,
-    CasCoreTicketCatalogConfiguration.class,
-    CasWebApplicationServiceFactoryConfiguration.class})
 @TestPropertySource(properties = {
     "cas.ticket.registry.redis.host=localhost",
     "cas.ticket.registry.redis.port=6320",
@@ -45,10 +24,8 @@ import redis.embedded.RedisServer;
     "cas.ticket.registry.redis.sentinel.node[1]=localhost:26740",
     "cas.ticket.registry.redis.sentinel.node[2]=localhost:26741"
 })
-@EnableTransactionManagement(proxyTargetClass = true)
-@EnableAspectJAutoProxy(proxyTargetClass = true)
 @DisabledIfContinuousIntegration
-public class SentinelEmbeddedTicketRegistryTests extends BaseTicketRegistryTests {
+public class SentinelEmbeddedTicketRegistryTests extends BaseRedisSentinelTicketRegistryTests {
 
     private static RedisServer REDIS_SERVER_1;
     private static RedisSentinel SENTINEL_SERVER_1;
@@ -56,10 +33,6 @@ public class SentinelEmbeddedTicketRegistryTests extends BaseTicketRegistryTests
     private static RedisSentinel SENTINEL_SERVER_2;
     private static RedisServer REDIS_SERVER_3;
     private static RedisSentinel SENTINEL_SERVER_3;
-
-    @Autowired
-    @Qualifier("ticketRegistry")
-    private TicketRegistry ticketRegistry;
 
     @BeforeAll
     public static void startRedis() throws Exception {
@@ -90,10 +63,5 @@ public class SentinelEmbeddedTicketRegistryTests extends BaseTicketRegistryTests
         REDIS_SERVER_3.stop();
         REDIS_SERVER_2.stop();
         REDIS_SERVER_1.stop();
-    }
-
-    @Override
-    public TicketRegistry getNewTicketRegistry() {
-        return this.ticketRegistry;
     }
 }
